### 2022-07-06 #4551 @editorialbot
[PRE REVIEW]: COINr: An R package for developing composite indicators
**Submitting author:** @bluefoxr (<a href="http://orcid.org/0000-0002-6467-4472">William Becker</a>)
**Repository:** https://github.com/bluefoxr/COINr
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
