### 2022-07-06 c6f8280
Updated dois and urls, cleaned up bib file
