+++ paper.bib
@@ -1,69 +1,3 @@
-@article{babaee2014combining,
-  title={Combining driving performance information in an index score: a simulated curve-taking experiment},
-  author={Babaee, Seddigheh and Shen, Yongjun and Hermans, Elke and Wets, Geert and Brijs, Tom and Ari{\"e}n, Caroline},
-  journal={Transportation Research Record},
-  volume={2434},
-  number={1},
-  pages={44--51},
-  year={2014},
-  publisher={SAGE Publications Sage CA: Los Angeles, CA}
-}
-
-@article{Bandura08,
-  title={A survey of composite indices measuring country performance: 2008 update},
-  author={Bandura, Romina},
-  journal={New York: United Nations Development Programme, Office of Development Studies (UNDP/ODS Working Paper)},
-  year={2008}
-}
-
-@TECHREPORT{Bandura11,
-  author = {Bandura, R.},
-  title = {Composite indicators and rankings: Inventory 2011},
-  institution = {Office of Development Studies, United Nations Development Programme (UNDP)},
-  year = {2011},
-  address = {New York}
-}
-
-@article{barclay2019problem,
-  title={The problem with composite indicators},
-  author={Barclay, Matthew and Dixon-Woods, Mary and Lyratzopoulos, Georgios},
-  journal={BMJ quality \& safety},
-  volume={28},
-  number={4},
-  pages={338--344},
-  year={2019},
-  publisher={BMJ Publishing Group Ltd}
-}
-
-@article{becker2017weights,
-  title={Weights and importance in composite indicators: Closing the gap},
-  author={Becker, William and Saisana, Michaela and Paruolo, Paolo and Vandecasteele, Ine},
-  journal={Ecological indicators},
-  volume={80},
-  pages={12--22},
-  year={2017},
-  publisher={Elsevier}
-}
-
-@book{becker2018_ASEM,
-	title = {Exploring {ASEM} sustainable connectivity: What brings {A}sia and {E}urope together?},
-	isbn = {978-92-79-99726-6},
-	shorttitle = {Exploring {ASEM} connectivity},
-	language = {English},
-	author = {Becker, W. and Dominguez Torreiro, M. and Neves, A. R. and Saisana, M. and Tacao Moura, C.},
-	year = {2018},
-	publisher = {Publications Office of the European Union, Luxembourg}
-}
-
-@article{becker2020wrapping,
-  title={Wrapping up the Europe 2020 strategy: A multidimensional indicator analysis},
-  author={Becker, William and Norl{\'e}n, Hedvig and Dijkstra, Lewis and Athanasoglou, Stergios},
-  journal={Environmental and Sustainability Indicators},
-  volume={8},
-  pages={100075},
-  year={2020},
-  publisher={Elsevier}
-}
 
 @article{becker2021exploring,
   title={Exploring the link between Asia and Europe connectivity and sustainable development},
@@ -72,15 +6,16 @@
   volume={3},
   pages={100045},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = "10.1016/j.resglo.2021.100045"
 }
 
-@book{becker2021_COINrManual,
-	title = {Composite Indicator Development and Analysis in R with COINr},
-	language = {English},
-	author = {Becker, W.},
-	year = {2021},
-	note = {\url{https://bluefoxr.github.io/COINrDoc/}}
+@manual{CIAOtool,
+      title        = "Composite Indicator Analysis and Optimization ({CIAO}) Tool, v.2",
+      author       = "Lindén, D. and Cinelli, M. and Spada, M. and Becker, W. and Burgherr, P.",
[…]
