+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: COINr
 Type: Package
 Title: Composite Indicator Construction and Analysis
-Version: 1.0.0
+Version: 1.0.0.9000
 Maintainer: William Becker <william.becker@bluefoxdata.eu>
 Description: A comprehensive high-level package, for composite indicator construction and analysis. It is a "development environment"
     for composite indicators and scoreboards, which includes utilities for construction (indicator selection, denomination, imputation,
