### 2022-07-08 619aa8b
Purse method fixed so that returns data when units are missing from some (but not all) coins
