### 2022-07-08 #4551 @jbytecode
👋👋👋 Dear @fartist 👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**COINr: An R package for developing composite indicators**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/4551).
