### 2022-07-11 #4551 @bluefoxr
@jbytecode I could suggest "paulrougieux" (not registered at JOSS I think but well familiar with R). I'll try to think of some others
