### 2022-07-13 #4551 @fartist
Me neither - have a move scheduled coming up, sorry.

Best,

On Wed, Jul 13, 2022 at 10:22 AM chrisaberson ***@***.***>
wrote:

> I won't be able to do this one - am scheduled to leave for a trip
>
