+++ paper.md
@@ -29,6 +29,7 @@ authors:
   affiliation: 2
 - name: Michaela Saisana
   affiliation: 2
+  orcid: 0000-0001-5361-0222
 bibliography: paper.bib
 ---
 
