### 2022-07-18 81d5995
Fix uCode retrieval for purse method when diff uCodes for diff coins
