### 2022-07-19 #4567 @paulrougieux
@jbytecode @bluefoxr ([William Becker](http://orcid.org/0000-0002-6467-4472)) is a former colleague from the Joint Research Centre, a large research organisation, but we were working in different research groups.
