### 2022-07-19 #4567 @danielskatz
Can you define a couple?
