### 2022-07-19 #4567 @bluefoxr
Hi @danielskatz I invited @paulrougieux  since we were short on reviewers. I worked in the European Commission's Joint Research Centre up until March 2020. Paul worked there as well, and still works there. He has a lot of experience in R and has actually used the COINr package (the subject of this paper). I am sure Paul would give an objective review, but if this is too much of a conflict of interest I understand and my apologies.
