### 2022-07-19 #4567 @danielskatz
Our COI rules use 4 years as the conflict in this type of situation, so the editor (@jbytecode) now need to decide whether this COI that we've found should be waived.
