### 2022-07-19 #4567 @danielskatz
@bluefoxr - You are correct.  I was misremembering the [JOSS guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), which say that the organizational conflict period is one year, while the collaboration conflict period is four years.  So there is no conflict here, and the review can proceed without concern.
