### 2022-07-19 bluefoxr/COINr#6 @bauer-alex
LICENSE should only be contained in a plain-text file
Issue based on the JOSS revision process (see [the corresponding issue](https://github.com/openjournals/joss-reviews/issues/4567)).

Please put the software license in one plain-text file _LICENSE_ only (similar to this [example package](https://github.com/bauer-alex/APCtools)) The file _LICENSE.md_ can then be deleted.
