### 2022-07-19 bluefoxr/COINr#7 @bauer-alex
Improve coverage of unit tests
Issue based on the JOSS revision process (see [the corresponding issue](https://github.com/openjournals/joss-reviews/issues/4567)).

Based on `devtools::test_coverage()`, the unit tests in the package currently only cover 25.80% of the relevant code. This number needs to be higher to accept the paper. JOSS doesn't state a specific required test coverage and of course reaching 100% is usually not realistic (as of a lot of specific subfunctionalities when making use of individual function arguments), but currently a lot of functions have no corresponding unit tests at all. Please include unit tests for all functions of the package.
