### 2022-07-19 bluefoxr/COINr#8 @bauer-alex
Include community guidelines in the readme file
Issue based on the JOSS revision process (see [the corresponding issue](https://github.com/openjournals/joss-reviews/issues/4567)).

Please include guidelines for third parties in the package's readme file, guiding people who wish to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
You can mainly copy the guidelines from some other package (for example [this one](https://github.com/bauer-alex/APCtools)).
