### 2022-07-19 bluefoxr/COINr#9 @bauer-alex
Empty examples in online vignettes 
Issue based on the JOSS revision process (see [the corresponding issue](https://github.com/openjournals/joss-reviews/issues/4567)).

Currently, the two _Examples_ subpages on [bluefoxr.github.io/COINr](https://bluefoxr.github.io/COINr/) are empty. Please fill them or delete these subpages for the moment. Also without them you currently have enough examples in the other vignettes.
