### 2022-07-19 bluefoxr/COINr#10 @bauer-alex
Add function deprecation notes in help pages
Issue based on the JOSS revision process (see [the corresponding issue](https://github.com/openjournals/joss-reviews/issues/4567)).

The package contains a lot of exported functions which are deprecated. Printing the `COINr:::defunct_message()` when such functions are called seems like a good idea. Only thing I also would to see is a short deprecation note (can be identical to the defunct message) in the help page (e.g. `?COINr::getPCA`) for full clarity.
