### 2022-07-19 #4567 @bauer-alex
@bluefoxr Seems like a very interesting paper! I especially appreciate the high effort you put into documenting all functionalities appropriately. I opened some technical and documentation issues in the [COINr repository](https://github.com/bluefoxr/COINr/issues). Once these are dealt with I will take a deeper look over the paper itself.

One additional question: In the GitHub repository you are the only contributor to the package. Could you clarify the contributions of your co-authors?
