### 2022-07-20 dd0f5a9
Licence moved to plain text file only
