### 2022-07-20 bluefoxr/COINr#6 @bluefoxr
Thanks, I have now moved the licence to the plain text file and deleted the markdown file
