### 2022-07-20 1d9dce1
Include collaboration and issues instructions as per JOSS guidelines
