### 2022-07-20 bluefoxr/COINr#8 @bluefoxr
Thanks, I fixed this now: see commit https://github.com/bluefoxr/COINr/commit/1d9dce183cb59494c1ba870e014b5ae85442c283
