### 2022-07-20 bluefoxr/COINr#7 @bluefoxr
Thanks, it's a fair point and I will improve this but this will take a while due to the number of functions. I'll put some updates here as I go along.
