### 2022-07-20 bluefoxr/COINr#9 @bluefoxr
Ok yes these are placeholders where I wanted to include examples where data is acquired through APIs and then processed through the package. However at the moment I don't have time to finish these so I will delete them as suggested. Will need to regenerate the pkgdown site so this will take a short while.
