### 2022-07-20 bluefoxr/COINr#10 @bluefoxr
OK good point, I will do this.
