### 2022-07-20 #4567 @bluefoxr
> @bluefoxr Seems like a very interesting package! I especially appreciate the high effort you put into documenting all functionalities appropriately. I opened some technical and documentation issues in the [COINr repository](https://github.com/bluefoxr/COINr/issues). Once these are dealt with I will take a deeper look over the paper itself.
>
> One additional question: In the GitHub repository you are the only contributor to the package. Could you clarify the contributions of your co-authors?

@bauer-alex Thanks very much for taking the time to look over the package. I will deal with all the issues you raised although as mentioned expanding the unit testing coverage will take a little while as this is quite a big task.

As regards the co-authors they are included due to a lot of helpful input in conceptualising the package, testing features and providing feedback.
