### 2022-07-21 43edd5a
Various functions for trend analysis
