### 2022-07-21 323565a
Add new funcs for trend analysis
