+++ NAMESPACE
@@ -40,6 +40,7 @@ S3method(qTreat,coin)
 S3method(qTreat,data.frame)
 S3method(qTreat,purse)
 export(Aggregate)
+export(CAGR)
 export(COINToolIn)
 export(COIN_to_coin)
 export(Denominate)
@@ -54,6 +55,7 @@ export(a_amean)
 export(a_copeland)
 export(a_gmean)
 export(a_hmean)
+export(approx_df)
 export(assemble)
 export(boxcox)
 export(build_ASEM)
@@ -105,6 +107,7 @@ export(get_results)
 export(get_sensitivity)
 export(get_stats)
 export(get_str_weak)
+export(get_trends)
 export(get_unit_summary)
 export(harMean)
 export(hicorrSP)
@@ -160,6 +163,7 @@ export(plot_scatter)
 export(plot_sensitivity)
 export(plot_uncertainty)
 export(plotframework)
+export(prc_change)
 export(qNormalise)
 export(qTreat)
 export(rankDF)
@@ -210,9 +214,11 @@ importFrom(openxlsx,write.xlsx)
 importFrom(readxl,cell_limits)
 importFrom(readxl,read_excel)
 importFrom(rlang,.data)
+importFrom(stats,approx)
 importFrom(stats,cor)
 importFrom(stats,cor.test)
 importFrom(stats,cov)
+importFrom(stats,lm)
 importFrom(stats,median)
 importFrom(stats,na.omit)
 importFrom(stats,optim)
