### 2022-07-26 bluefoxr/COINr#13 @paulrougieux
paper.md language suggestions
Comments related to the JOSS review https://github.com/openjournals/joss-reviews/issues/4567

Line numbers refer to the version of paper.md on July 20 2022 at commit 1d9dce183cb59494c1ba870e014b5ae85442c283.

Line 40
> "fast development" ... " quick exploration"

The adjectives fast and quick can be removed, unless a documented comparison with other
tools can justify using these adjectives.

Line 50

> In Python there is a library called CIF which gives

Language suggestion:

The Python library CIF gives …

Line 54
> "full global sensitivity analysis"

This expression is repeated lines 40 and line 83. How does this differ from a
sensitivity analysis that would not be “full” and “global”? If this is a specific type
of sensitivity analysis, maybe this could be defined here?
