### 2022-07-26 bluefoxr/COINr#14 @paulrougieux
paper.md explain why the "coin" and "purse" objects are necessary
Related to the JOSS review https://github.com/openjournals/joss-reviews/issues/4567

The line number refers to the version of paper.md on July 20 2022 at commit 1d9dce183.

Line 54

> "COINr wraps all composite indicator data, analysis and methodological choices into a
> single S3 class object called a “coin”. This enables a neat and structured
> environment, simplifies the syntax of functions ..." COINr also supports time-indexed
> data, represented by the "purse" class (a time-indexed collection of coins)."

Can you provide more details on the structure of those objects?
What are the advantage of the coin and purse objects over standard data frames,
potentially with list columns (à la dplyr, tidyr, purrr) as described by Hadley Wickham in [many models](https://r4ds.had.co.nz/many-models.html)?

Is this new coin structure required because of the nature of the data?
