### 2022-07-26 bluefoxr/COINr#15 @paulrougieux
paper.md illustrate the problem the software is designed to solve
Related to the JOSS review https://github.com/openjournals/joss-reviews/issues/4567

For example in issue #6  Alex Bauer referenced another package called APCtools. The associated [JOSS paper](https://joss.theoj.org/papers/10.21105/joss.04056) illustrates the main features of the package with several figures (density matrix faceted plots and heat maps). Those figures are of great help to provide an overview of what the package can do.

There is plenty of material in the COINr vignettes to provide an illustration of the software's main features, a selection of code examples and figures could be made available in the paper. Some examples from the [normalization](https://bluefoxr.github.io/COINr/articles/normalise.html) and obviously the [visualization](https://bluefoxr.github.io/COINr/articles/visualisation.html) vignette could be reused.
