### 2022-07-26 bluefoxr/COINr#19 @paulrougieux
Is there an inconsistency in the normalized values?
In the example dataset, I don't understand how the normalised values 86, 65 and 100 can be created from the raw values 6.2, 15 and 0.6, Is there an inconsistency?

```
library(COINr)
coin <- build_example_coin(quietly = TRUE)
cr <- coin$Data$Raw
cn <- coin$Data$Normalised
cr[cr$uCode %in% c("SWE", "THA", "VNM"), "NEET"]
# [1]  6.2 15.0  0.6

cn[cn$uCode %in% c("SWE", "THA", "VNM"), "NEET"]
# [1]  86.50602  65.30120 100.00000
```
Related to https://github.com/openjournals/joss-reviews/issues/4567
