### 2022-08-09 #4567 @bluefoxr
Dear @jbytecode my apologies for the slow reply, I have been away. I have addressed some initial comments previously and I see that there are a number of extra comments to address. Most of these can be dealt with fairly quickly but some, such as the unit testing coverage, will need quite a bit of time to improve. I will see what progress I can make in the next week, after which I will again be away until the beginning of September. Would it be reasonable to ask until end of Sep to deal with this? I am mainly thinking of the time needed to code up all the unit tests. Thanks.
