### 2022-08-09 bluefoxr/COINr#13 @bluefoxr
@paulrougieux thanks for these comments, I will make the suggested changes
