### 2022-08-09 bluefoxr/COINr#14 @bluefoxr
@paulrougieux the coin class is defined which is a full representation of a composite indicator - see here: https://bluefoxr.github.io/COINr/articles/coins.html#what-is-a-coin

It includes a lot more than just the data, including also the index structure (a hierarchical structure), sets of weights, methodological specifications (which allow coins to be regenerated), metadata for both units and indicators, and some frequently used parameters which save processing time in functions. This is a lot more than can be wrapped in a data frame, and is not rectangular data.

To see the contents of a coin you can call its `print()` method:

```
library(COINr)
# build example coin
ASEM <- build_example_coin()
ASEM
```

for a more detailed exploration try e.g.:

```
str(ASEM, max.level = 2)
```

The "purse" class is indeed a data frame with a list column in which each entry is a coin. So it is a way of indexing many coins by time.

There are several reasons for defining these classes. First that building functions (e.g. `Treat()` and others) have methods for numeric vectors and data frames as well as coins and purses. So using method dispatch seems a logical way to keep these functions flexible for users.

Then as mentioned, they keep all the data and metadata in a neat order, and allow copies and regeneration, which would otherwise not be possible (or at least far messier): https://bluefoxr.github.io/COINr/articles/adjustments.html
