### 2022-08-09 bluefoxr/COINr#14 @bluefoxr
@paulrougieux do you think some further explanation is necessary in the paper? I was trying to keep it brief as per the guidelines.
