### 2022-08-10 bluefoxr/COINr#14 @paulrougieux
Since coins are the central objects of your package, I think it is important to briefly describe how they are structured. Based on your message:

> "It includes [...] the index structure (a hierarchical structure), sets of weights, methodological specifications (which allow coins to be regenerated), metadata for both units and indicators, and some frequently used parameters which save processing time in functions. This is a lot more than can be wrapped in a data frame, and is not rectangular data."

The paper could explain the nature of coins in a succinct way and refer to the following page for the full description https://bluefoxr.github.io/COINr/articles/coins.html#what-is-a-coin
