### 2022-08-10 bluefoxr/COINr#14 @bluefoxr
Yes ok then I will add this to the paper. Thanks
