### 2022-08-11 a1f2c6b
Add x/y output for time series
