### 2022-08-17 71ba9f7
Fix aggregate.data.frame so can deal with all NAs
