### 2022-08-17 e47c49e
Purse method for export to excel
