+++ R/export.R
@@ -4,9 +4,10 @@
 #' frame on a separate tab. Tabs are named according to the position in the coin object. You can write other
 #' data frames by simply attaching them to the coin object somewhere.
 #'
-#' @param coin A coin class object
+#' @param x A coin class object
 #' @param fname The file name/path to write to, as a character string
 #' @param include_log Logical: if `TRUE`, also writes data frames from the `.$Log` list inside the coin.
+#' @param ... arguments passed to or from other methods.
 #'
 #' @importFrom openxlsx write.xlsx
 #'
@@ -30,7 +31,7 @@
 #' @return .xlsx file at specified path
 #'
 #' @export
-export_to_excel <- function(coin, fname = "coin_export.xlsx", include_log = FALSE){
+export_to_excel.coin <- function(x, fname = "coin_export.xlsx", include_log = FALSE){
 
   check_coin_input(coin)
 
@@ -65,4 +66,116 @@ export_to_excel <- function(coin, fname = "coin_export.xlsx", include_log = FALS
   # write to excel
   openxlsx::write.xlsx(coinwrite, file = fname, colNames = TRUE)
 
+}
+
+
+#' Export a purse to Excel
+#'
+#' Exports the contents of the purse to Excel. This is similar to the coin method [export_to_excel.coin()],
+#' but combines data sets from various time points. It also selectively writes metadata since this may be
+#' spread across multiple coins.
+#'
+#' @param x A purse class object
+#' @param fname The file name/path to write to, as a character string
+#' @param include_log Logical: if `TRUE`, also writes data frames from the `.$Log` list inside the coin.
+#' @param ... arguments passed to or from other methods.
+#'
+#' @importFrom openxlsx write.xlsx
+#'
+#' @examples
+#' #
+#'
+#' @return .xlsx file at specified path
+#'
+#' @export
+export_to_excel.purse <- function(x, fname = "coin_export.xlsx", include_log = FALSE){
+
+
+  # Prep --------------------------------------------------------------------
+
+  check_purse_input(x)
+
+  # function to stop tab names exceeding 31 characters, avoiding errors.
+  trunc_str <- function(x){
+    if(nchar(x) > 31){
+      xnew <- substr(x, 1, 31)
+      warning("Truncated tab name (", x, ") because exceeds Excel length limit (31 characters).")
+    } else {
+      xnew <- x
+    }
+    xnew
+  }
+
+
+  # Extract coin ------------------------------------------------------------
+  # For the purse method, we take the first coin and modify it:
+  # - merge dsets over time
+  # - small other tweaks
+  # - then write the coin
+
+  # Extract first coin in the purse
+  coin <- x$coin[[1]]
+
+  # get names of data sets first
+  dset_names <- names(coin$Data)
+
+  # get data sets with all metadata attached
+  dsets <- lapply(dset_names, function(dset){
+    get_dset(x, dset, also_get = "all")
+  })
+  names(dsets) <- dset_names
+
+  # replace data list with full dsets over time
+  coin$Data <- dsets
+
+  # I have to reconstruct the full uMeta because coins have different numbers of units
+  icodes <- coin$Meta$Ind$iCode[coin$Meta$Ind$Type %in% c("Indicator", "Aggregate")]
+  uMeta <- unique(dsets[[1]][!names(dsets[[1]]) %in% c(icodes, "Time")])
+  coin$Meta$Unit <- uMeta
+
+  # remove log if necessary
+  if(!include_log){
+    coin$Log <- NULL
+  }
+
+  # Assemble into list ------------------------------------------------------
+
[…]
