+++ man/export_to_excel.purse.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/export.R
+\name{export_to_excel.purse}
+\alias{export_to_excel.purse}
+\title{Export a purse to Excel}
+\usage{
+\method{export_to_excel}{purse}(x, fname = "coin_export.xlsx", include_log = FALSE)
+}
+\arguments{
+\item{x}{A purse class object}
+
+\item{fname}{The file name/path to write to, as a character string}
+
+\item{include_log}{Logical: if \code{TRUE}, also writes data frames from the \code{.$Log} list inside the coin.}
+
+\item{...}{arguments passed to or from other methods.}
+}
+\value{
+.xlsx file at specified path
+}
+\description{
+Exports the contents of the purse to Excel. This is similar to the coin method \code{\link[=export_to_excel.coin]{export_to_excel.coin()}},
+but combines data sets from various time points. It also selectively writes metadata since this may be
+spread across multiple coins.
+}
+\examples{
+#
+
+}
