+++ NAMESPACE
@@ -23,6 +23,8 @@ S3method(Treat,coin)
 S3method(Treat,data.frame)
 S3method(Treat,numeric)
 S3method(Treat,purse)
+S3method(export_to_excel,coin)
+S3method(export_to_excel,purse)
 S3method(get_data,coin)
 S3method(get_data,purse)
 S3method(get_data_avail,coin)
