### 2022-08-17 fa26eec
Option to adjust trends by directions or not
