### 2022-08-17 49135a7
News post merge of trend branch
