### 2022-08-17 b687bfa
fix for cmd check
