### 2022-08-29 3bdb48e
Update doc for aggregation by level
