+++ man/Aggregate.coin.Rd
@@ -22,11 +22,16 @@
 
 \item{dset}{The name of the data set to apply the function to, which should be accessible in \code{.$Data}.}
 
-\item{f_ag}{The name of an aggregation function, a string}
+\item{f_ag}{The name of an aggregation function, a string. This can either be a single string naming
+a function to use for all aggregation levels, or else a character vector of function names of length \code{n-1}, where \code{n} is
+the number of levels in the index structure. In this latter case, a different aggregation function may be used for each level
+in the index: the first in the vector will be used to aggregate from Level 1 to Level 2, the second from Level 2 to Level 3, and
+so on.}
 
 \item{w}{An optional data frame of weights. If \code{f_ag} does not require accept weights, set to \code{"none"}.}
 
-\item{f_ag_para}{Optional parameters to pass to \code{f_ag}, other than \code{x} and \code{w}}
+\item{f_ag_para}{Optional parameters to pass to \code{f_ag}, other than \code{x} and \code{w}. As with \code{f_ag}, this can specified to have different
+parameters for each aggregation level by specifying as a nested list of length \code{n-1}.}
 
 \item{dat_thresh}{An optional data availability threshold, specified as a number between 0 and 1. If a row
 within an aggregation group has data availability lower than this threshold, the aggregated value for that row will be
