### 2022-09-07 bec3f55
include new funcs in function index on pkgdown
