### 2022-09-07 204c84b
include new funcs in function index on pkgdown
