+++ _pkgdown.yml
@@ -96,6 +96,7 @@ reference:
   - get_sensitivity
   - get_stats
   - get_str_weak
+  - get_trends
   - get_unit_summary
   - remove_elements
 
@@ -134,11 +135,13 @@ reference:
   - Aggregate.coin
   - Aggregate.data.frame
   - Aggregate.purse
+  - approx_df
   - a_amean
   - a_copeland
   - a_gmean
   - a_hmean
   - boxcox
+  - CAGR
   - check_iData
   - check_iMeta
   - check_SkewKurt
@@ -147,6 +150,8 @@ reference:
   - Denominate.coin
   - Denominate.data.frame
   - Denominate.purse
+  - export_to_excel.coin
+  - export_to_excel.purse
   - get_data.coin
   - get_data.purse
   - get_data_avail.coin
@@ -191,6 +196,7 @@ reference:
   - n_scaled
   - n_zscore
   - outrankMatrix
+  - prc_change
   - print.coin
   - print.purse
   - qNormalise.coin
