### 2022-09-07 91b7733
Remove dead examples
