+++ vignettes/panel_data.Rmd
@@ -1,17 +1,0 @@
----
-title: "Panel Data"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Panel Data}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-This vignette will walk through analysis of panel data in COINr. I will finish this soonish.
