### 2022-09-14 f98fc6d
Defunct messages in help pages of functions (issue #10)
