+++ R/plot_dist.R
@@ -6,6 +6,8 @@
 #' This function uses ggplot2 to generate plots, so the plot can be further manipulated using ggplot2 commands.
 #' See `vignette("visualisation`) for more details on plotting.
 #'
+#' This function replaces the now-defunct `plotIndDist()` from COINr < v1.0.
+#'
 #' @param coin The coin object, or a data frame of indicator data
 #' @param dset The name of the data set to apply the function to, which should be accessible in `.$Data`.
 #' @param iCodes Indicator code(s) to plot. See details.
@@ -126,6 +128,8 @@ plot_dist <- function(coin, dset, iCodes, ..., type = "Box", normalise = FALSE,
 #'
 #' This function uses ggplot2 to generate plots, so the plot can be further manipulated using ggplot2 commands.
 #' See `vignette("visualisation`) for more details on plotting.
+#'
+#' This function replaces the now-defunct `plotIndDot()` from COINr < v1.0.
 #'
 #' @param coin The coin
 #' @param dset The name of the data set to apply the function to, which should be accessible in `.$Data`.
