### 2022-09-14 bluefoxr/COINr#10 @bluefoxr
I have now added deprecation notes to all the relevant functions - see commit here: https://github.com/bluefoxr/COINr/commit/f98fc6df5a9fd08ce67caf11cc41f275fd8b8701

I'll close this issue now, because I presume it is fixed - if not, please open a separate issue. Thanks.
