### 2022-09-14 a1340a0
Small edits referring to issue #13
