+++ paper.bib
@@ -110,6 +110,27 @@
   doi = "10.1016/j.envsoft.2021.105208"
 }
 
+@article{saisana2005uncertainty,
+  title={Uncertainty and sensitivity analysis techniques as tools for the quality assessment of composite indicators},
+  author={Saisana, Michaela and Saltelli, Andrea and Tarantola, Stefano},
+  journal={Journal of the Royal Statistical Society: Series A (Statistics in Society)},
+  volume={168},
+  number={2},
+  pages={307--323},
+  year={2005},
+  publisher={Wiley Online Library}
+}
+
+@article{saltelli2019so,
+  title={Why so many published sensitivity analyses are false: A systematic review of sensitivity analysis practices},
+  author={Saltelli, Andrea and Aleksankina, Ksenia and Becker, William and Fennell, Pamela and Ferretti, Federico and Holst, Niels and Li, Sushan and Wu, Qiongli},
+  journal={Environmental modelling \& software},
+  volume={114},
+  pages={29--39},
+  year={2019},
+  publisher={Elsevier}
+}
+
 @article{stefana2021composite,
   title={Composite indicators to measure quality of working life in Europe: a systematic review},
   author={Stefana, Elena and Marciano, Filippo and Rossi, Diana and Cocca, Paola and Tomasoni, Giuseppe},
