### 2022-09-14 bluefoxr/COINr#13 @bluefoxr
Hi, I've made edits to the paper to fix the issues: see commit https://github.com/bluefoxr/COINr/commit/a1340a02f488d45a6ccaee1182d025d8f7986493

For the global sensitivity analysis, I have left a reference because my understanding of JOSS papers is that they should be quite brief, so I would prefer to leave the explanation elsewhere. See if you think that is OK then I will close the issue or make some further edits. Thanks.
