### 2022-09-14 ded8997
More explanation on what a coin is, addressing issue #14
