### 2022-09-14 bluefoxr/COINr#14 @bluefoxr
See commit https://github.com/bluefoxr/COINr/commit/ded8997f83b55628906dfdcd85771d9ebad9552c

Please let me know if the edits sufficiently address the issue. Thanks,
