### 2022-09-14 bluefoxr/COINr#14 @paulrougieux
The update
> "A coin is a structured list including:
> * Indicator data sets for each processing step (e.g. imputation, normalisation, etc.)
> * Metadata pertaining to indicators and units (e.g. names and weights, but also the hierarchical structure of the index)
> * A record of the COINr functions applied in constructing the coin"

provides good details on how the coin objects are structured. OK from my side.
