### 2022-09-14 bluefoxr/COINr#13 @paulrougieux
The reference is ok.
