### 2022-09-14 bluefoxr/COINr#19 @bluefoxr
Hi, this is because the example uses "min-max" normalisation, which just re-scales the data to lie on the [0,100] interval.

Following your code above, try:
```
plot(cr$NEET, cn$NEET)
```
![image](https://user-images.githubusercontent.com/72880276/190203914-da7df50b-a0c3-424d-8f23-0f186066d40f.png)

You can see that the normalised data lies between [0,100] and is also reversed because NEET is a "negative" indicator, i.e. higher values are negatively associated with the concept (in this case social sustainability).

Does that clear up the doubt?
