### 2022-09-14 bluefoxr/COINr#19 @bluefoxr
See also [minmax()](https://bluefoxr.github.io/COINr/reference/n_minmax.html) which is driving this operation
