### 2022-09-14 27cfa02
Fix formatting RE issue #18
