### 2022-09-14 bluefoxr/COINr#19 @paulrougieux
```

> cr$NEET
 [1]  8.7  6.5  9.3 27.4 15.3 17.2  6.5   NA 16.1  6.3  6.3  7.0 13.3  9.4  9.4
[16] 11.5 10.3 15.3 15.4 11.0 21.5 27.9 10.9 20.0  3.5  9.5 12.7   NA 42.1  9.2
[31]  5.9 10.3  8.0 17.4 19.8 12.8  4.0  4.6 11.8 30.4 21.7  9.5  9.3 15.2 12.4
[46]  4.0 12.1  6.5  6.2 15.0  0.6
> n_minmax(-cr$NEET)
 [1]  80.48193  85.78313  79.03614  35.42169  64.57831  60.00000  85.78313
 [8]        NA  62.65060  86.26506  86.26506  84.57831  69.39759  78.79518
[15]  78.79518  73.73494  76.62651  64.57831  64.33735  74.93976  49.63855
[22]  34.21687  75.18072  53.25301  93.01205  78.55422  70.84337        NA
[29]   0.00000  79.27711  87.22892  76.62651  82.16867  59.51807  53.73494
[36]  70.60241  91.80723  90.36145  73.01205  28.19277  49.15663  78.55422
[43]  79.03614  64.81928  71.56627  91.80723  72.28916  85.78313  86.50602
[50]  65.30120 100.00000
> cn$NEET
 [1]  80.48193  85.78313  79.03614  35.42169  64.57831  60.00000  85.78313
 [8]  61.31896  62.65060  86.26506  86.26506  84.57831  69.39759  78.79518
[15]  78.79518  73.73494  76.62651  64.57831  64.33735  74.93976  49.63855
[22]  34.21687  75.18072  53.25301  93.01205  78.55422  70.84337  61.31896
[29]   0.00000  79.27711  87.22892  76.62651  82.16867  59.51807  53.73494
[36]  70.60241  91.80723  90.36145  73.01205  28.19277  49.15663  78.55422
[43]  79.03614  64.81928  71.56627  91.80723  72.28916  85.78313  86.50602
[50]  65.30120 100.00000
>

```

Clarifies the issue thanks.
