### 2022-09-14 b9a7721
Added more explanation of features plus very short demo for issue #15
