+++ paper.md
@@ -5,7 +5,7 @@ tags:
 - composite indicators
 - indicators
 - policy
-date: 6 July 2022
+date: "14 September 2022"
 affiliations:
 - name: Freelance consultant, Ispra, Italy
   index: 1
@@ -15,13 +15,13 @@ affiliations:
   index: 3
 authors:
 - name: William Becker
-  orcid: 0000-0002-6467-4472
+  orcid: "0000-0002-6467-4472"
   corresponding: yes
   affiliation: 1
 - name: Giulio Caperna
   affiliation: 2
 - name: Maria Del Sorbo
-  orcid: 0000-0001-8717-0202
+  orcid: "0000-0001-8717-0202"
   affiliation: 3
 - name: Hedvig Norlén
   affiliation: 1
@@ -29,7 +29,7 @@ authors:
   affiliation: 2
 - name: Michaela Saisana
   affiliation: 2
-  orcid: 0000-0001-5361-0222
+  orcid: "0000-0001-5361-0222"
 bibliography: paper.bib
 ---
 
@@ -67,34 +67,97 @@ In short, **COINr** aims to be a flexible, fast and comprehensive development en
 
 # Features
 
-**COINr** is extensively documented with many vignettes and examples, all of which can be easily browsed at its **pkgdown** [website](https://bluefoxr.github.io/COINr/). Here, a brief overview of features is given.
+**COINr** is extensively documented with many vignettes and examples, all of which can be easily browsed at its **pkgdown** [website](https://bluefoxr.github.io/COINr/). Here, a brief overview is given.
+
+Primarily, COINr is used for building composite indicators: in practice this would usually involve assembling a set of indicators (usually from different sources) and accompanying metadata, and assembling them into a data frame that can be read by COINr to build a "coin" (see [vignette](https://bluefoxr.github.io/COINr/articles/coins.html)). After that, the composite scores are calculated by operating on the coin using any of the "building functions", which specify *which* methodological steps to apply, and *how* to apply them.
+
+The full process of building a composite indicator is too lengthy to describe in this brief paper. Instead, we simply give a very short example. We use the built-in "ASEM" data set which comprises two data frames (one of indicator data, and the other of metadata) that are formatted such that they can be recognised by COINr to build a coin. To build a coin, we call `new_coin()`:
+
+```
+# load COINr
+library(COINr)
+
+# build a coin with example data set
+coin <- new_coin(iData = ASEM_iData, iMeta = ASEM_iMeta)
+```
+
+To see how these data frames are formatted, use e.g. `str(ASEM_iData)` or `View(ASEM_iData)` and see the "coins" [vignette](https://bluefoxr.github.io/COINr/articles/coins.html).
+
+In the most simple case, we could build a composite indicator by [normalising](https://bluefoxr.github.io/COINr/articles/normalise.html) the indicators (bringing them onto a common scale), and [aggregating](https://bluefoxr.github.io/COINr/articles/aggregate.html) them (using weighted averages to calculate index scores). This can be done in COINr using the `Normalise()` and `Aggregate()` functions respectively:
+
+```
+# normalise (scale) each indicator onto [0, 100] interval
+coin <- qNormalise(coin, dset = "Raw", f_n = "n_minmax",
+                   f_n_para = list(l_u = c(0, 100)))
+                   
+# aggregate using weighted arithmetic mean
+# (note weights are input in data frames when calling new_coin() )
+coin <- Aggregate(coin, dset = "Normalised", f_ag = "a_amean") 
+```
[…]
