### 2022-09-14 bluefoxr/COINr#15 @bluefoxr
hi @paulrougieux please see the amendments in commit https://github.com/bluefoxr/COINr/commit/b9a7721f11433565689b43f4c2a61aca4741cc9e

I have tried to strike a balance between adding a little more detail for the purposes of the paper, and respecting the requirement of brevity of the journal. The package is well documented at its website, so I have linked out as much as possible. Please see what you think and let me know if you see the need for adding more or changing something. Thanks.
