### 2022-09-14 bluefoxr/COINr#15 @bluefoxr
Compiled paper is here btw: https://github.com/bluefoxr/COINr/actions/runs/3055282246
