### 2022-09-15 bluefoxr/COINr#15 @paulrougieux
There is an issue with the current version of figure 1
![Screenshot from 2022-09-15 10-48-38](https://user-images.githubusercontent.com/5647397/190359517-9c57d5ed-02cf-4720-8451-55c92163127d.png)
