### 2022-09-15 bluefoxr/COINr#15 @paulrougieux
The 2 figures have the same caption in the pdf version of the paper available [here](https://github.com/bluefoxr/COINr/actions/runs/3055306832).
