### 2022-09-15 29f0b93
Fixed captions plus ref RE issue #15
