### 2022-09-15 bluefoxr/COINr#15 @bluefoxr
Sorry, I copied the code from another paper and forgot to change the captions. I have now fixed this, see https://github.com/bluefoxr/COINr/commit/29f0b934ce8f0024d9ee1abe0c2c8ec21582ea9a
