### 2022-09-22 81d5ede
add COINr6 due to tests
