### 2022-09-22 c3c5b1d
remove all defunct function redirects
