+++ NAMESPACE
@@ -43,7 +43,6 @@ S3method(qTreat,data.frame)
 S3method(qTreat,purse)
 export(Aggregate)
 export(CAGR)
-export(COINToolIn)
 export(COIN_to_coin)
 export(Denominate)
 export(Impute)
@@ -58,41 +57,17 @@ export(a_copeland)
 export(a_gmean)
 export(a_hmean)
 export(approx_df)
-export(assemble)
 export(boxcox)
-export(build_ASEM)
 export(build_example_coin)
 export(build_example_purse)
 export(change_ind)
-export(checkData)
 export(check_SkewKurt)
 export(check_iData)
 export(check_iMeta)
-export(coin2Excel)
-export(coin_win)
-export(colourTable)
-export(compTable)
-export(compTableMulti)
-export(compareDF)
 export(compare_coins)
 export(compare_coins_multi)
 export(compare_df)
-export(copeland)
-export(corrweightscat)
-export(effectiveWeight)
 export(export_to_excel)
-export(extractYear)
-export(geoMean)
-export(geoMean_rescaled)
-export(getCorr)
-export(getCronbach)
-export(getIn)
-export(getPCA)
-export(getResults)
-export(getStats)
-export(getStrengthNWeak)
-export(getUnitReport)
-export(getUnitSummary)
 export(get_PCA)
 export(get_corr)
 export(get_corr_flags)
@@ -111,30 +86,18 @@ export(get_stats)
 export(get_str_weak)
 export(get_trends)
 export(get_unit_summary)
-export(harMean)
-export(hicorrSP)
 export(i_mean)
 export(i_mean_grp)
 export(i_median)
 export(i_median_grp)
 export(import_coin_tool)
 export(impute_panel)
-export(indChange)
-export(indDash)
-export(iplotBar)
-export(iplotCorr)
-export(iplotIndDist)
-export(iplotIndDist2)
-export(iplotMap)
-export(iplotRadar)
-export(iplotTable)
 export(is.coin)
 export(is.purse)
 export(kurt)
 export(log_CT)
 export(log_CT_orig)
 export(log_GII)
-export(loggish)
 export(n_borda)
 export(n_dist2max)
 export(n_dist2ref)
@@ -146,16 +109,9 @@ export(n_prank)
 export(n_rank)
 export(n_scaled)
 export(n_zscore)
-export(names2Codes)
 export(names_to_codes)
 export(new_coin)
-export(noisyWeights)
 export(outrankMatrix)
-export(plotCorr)
-export(plotIndDist)
-export(plotIndDot)
-export(plotSA)
-export(plotSARanks)
 export(plot_bar)
 export(plot_corr)
 export(plot_dist)
@@ -164,24 +120,15 @@ export(plot_framework)
 export(plot_scatter)
 export(plot_sensitivity)
 export(plot_uncertainty)
-export(plotframework)
 export(prc_change)
 export(qNormalise)
 export(qTreat)
-export(rankDF)
 export(rank_df)
-export(removeElements)
 export(remove_elements)
-export(replaceDF)
 export(replace_df)
-export(resultsDash)
-export(rew8r)
-export(roundDF)
 export(round_df)
-export(sensitivity)
 export(signif_df)
 export(skew)
-export(weightOpt)
 export(winsorise)
 importFrom(ggplot2,aes)
 importFrom(ggplot2,coord_flip)
