### 2022-09-22 55e4c77
doc plus tiny fix in order of checks
