### 2022-09-22 cc3fcdc
fix bugs in outranking matrix and copeland
