### 2022-09-22 bd2fa96
fix unwanted time col
