+++ R/correlations.R
@@ -188,7 +188,8 @@ get_corr_flags <- function(coin, dset, cor_thresh = 0.9, thresh_type = "high", c
 #'
 #' Helper function for getting correlations between indicators and aggregates. This retrieves subsets of correlation
 #' matrices between different aggregation levels, in different formats. By default, it will return a
-#' long-form data frame, unless `make_long = FALSE`.
+#' long-form data frame, unless `make_long = FALSE`. By default, any correlations with a p-value less than 0.05 are
+#' replaced with `NA`. See `pval` argument to adjust this.
 #'
 #' This function allows you to obtain correlations between any subset of indicators or aggregates, from
 #' any data set present in a coin. Indicator selection is performed using [get_data()]. Two different
@@ -221,7 +222,9 @@ get_corr_flags <- function(coin, dset, cor_thresh = 0.9, thresh_type = "high", c
 #' Defaults to indicator level.
 #' @param ... Further arguments to be passed to [get_data()] (`uCodes` and `use_group`).
 #' @param cortype The type of correlation to calculate, either `"pearson"`, `"spearman"`, or `"kendall"`.
-#' @param withparent If `TRUE`, and `aglev[1] != aglev[2]`, will only return correlations of each row with its parent.
+#' @param withparent If `TRUE`, and `aglev[1] != aglev[2]`, will only return correlations of each row with its parent. Alternatively, if
+#' `withparent = "family"`, will return correlations with parents, grandparents etc, up to the highest level. In both cases the data set
+#' must be aggregated for this to work.
 #' @param grouplev The aggregation level to group correlations by if `aglev[1] == aglev[2]`. Requires that
 #' `make_long = TRUE`.
 #' @param pval The significance level for including correlations. Correlations with \eqn{p > pval} will be returned as `NA`.
