### 2022-09-22 fd40216
Removed all defunct function redirects
