+++ R/defunct.R
@@ -1,529 +1,529 @@
-# redirects for defunct functions
-
-defunct_message <-  function(){
-  message("COINr syntax and functionality has significantly changed. See vignette('v1') for details.")
-}
-
-COINrX_message <-  function(){
-  stop("This function has been removed from the main COINr package but is now available in COINrX.
-       See vignette('v1') for details.",  call. = FALSE)
-}
-
-# BoxCox
+# # redirects for defunct functions
 #
-# @param ... deprecated parameter
-# @export
-BoxCox <- function(...){
-  defunct_message()
-  .Defunct("boxcox")
-}
-
-#' COINToolIn
-#'
-#' @param ... deprecated parameter
-#' @export
-COINToolIn <- function(...){
-  defunct_message()
-  .Defunct("import_coin_tool")
-}
-
-#' assemble
-#'
-#' @param ... deprecated parameter
-#' @export
-assemble <- function(...){
-  defunct_message()
-  .Defunct("new_coin")
-}
-
-#' build_ASEM
-#'
-#' @param ... deprecated parameter
-#' @export
-build_ASEM <- function(...){
-  defunct_message()
-  .Defunct("build_example_coin")
-}
-
-#' checkData
-#'
-#' @param ... deprecated parameter
-#' @export
-checkData <- function(...){
-  defunct_message()
-  .Defunct("Screen")
-}
-
-#' coin2Excel
-#'
-#' @param ... deprecated parameter
-#' @export
-coin2Excel <- function(...){
-  defunct_message()
-  .Defunct("export_to_excel")
-}
-
-#' coin_win
-#'
-#' @param ... deprecated parameter
-#' @export
-coin_win <- function(...){
-  defunct_message()
-  .Defunct("winsorise")
-}
-
-#' colourTable
-#'
-#' @param ... deprecated parameter
-#' @export
-colourTable <- function(...){
-  COINrX_message()
-}
-
-#' compTable
-#'
-#' @param ... deprecated parameter
-#' @export
-compTable <- function(...){
-  defunct_message()
-  .Defunct("compare_coins")
-}
-
-#' compTableMulti
-#'
-#' @param ... deprecated parameter
-#' @export
-compTableMulti <- function(...){
-  defunct_message()
-  .Defunct("compare_coins_multi")
-}
-
-#' compareDF
-#'
-#' @param ... deprecated parameter
-#' @export
-compareDF <- function(...){
-  defunct_message()
-  .Defunct("compare_df")
-}
-
-#' copeland
-#'
-#' @param ... deprecated parameter
-#' @export
-copeland <- function(...){
-  defunct_message()
-  .Defunct("a_copeland")
-}
-
-#' corrweightscat
-#'
-#' @param ... deprecated parameter
-#' @export
-corrweightscat <- function(...){
-  COINrX_message()
-}
-
-# denominate
+# defunct_message <-  function(){
+#   message("COINr syntax and functionality has significantly changed. See vignette('v1') for details.")
+# }
 #
-# @param ... deprecated parameter
-# @export
-denominate <- function(...){
-  defunct_message()
-  .Defunct("Denominate")
-}
-
-#' effectiveWeight
-#'
-#' @param ... deprecated parameter
-#' @export
-effectiveWeight <- function(...){
-  defunct_message()
-  .Defunct("get_eff_weights")
-}
-
-#' extractYear
-#'
-#' @param ... deprecated parameter
-#' @export
-extractYear <- function(...){
-  defunct_message()
[…]
