### 2022-09-22 e314e66
fix bug on imputing panel data, also bug when not specified imputation type
