### 2022-09-22 a0e8c7a
fix issue where wrong iData was written to log when indicators excluded
