### 2022-09-22 f1afbff
fix issue when not use_latest, plus doc
