### 2022-09-22 cc7487a
fix incorrect writing of metadata
