### 2022-09-22 8fe1df8
remove package startup message
