+++ R/zzz.R
@@ -1,4 +1,4 @@
-.onAttach <- function(...) {
-  packageStartupMessage("COINr syntax has significantly changed. See vignette('v1') for details.
-This message will be removed in future updates.")
-}
+# .onAttach <- function(...) {
+#   packageStartupMessage("COINr syntax has significantly changed. See vignette('v1') for details.
+# This message will be removed in future updates.")
+# }
