### 2022-09-22 50d29c4
remove defunct function doc
