+++ man/iplotMap.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{iplotMap}
-\alias{iplotMap}
-\title{iplotMap}
-\usage{
-iplotMap(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-iplotMap
-}
