+++ man/names2Codes.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{names2Codes}
-\alias{names2Codes}
-\title{names2Codes}
-\usage{
-names2Codes(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-names2Codes
-}
