+++ man/harMean.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{harMean}
-\alias{harMean}
-\title{harMean}
-\usage{
-harMean(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-harMean
-}
