+++ man/replaceDF.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{replaceDF}
-\alias{replaceDF}
-\title{replaceDF}
-\usage{
-replaceDF(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-replaceDF
-}
