+++ man/getStats.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{getStats}
-\alias{getStats}
-\title{getStats}
-\usage{
-getStats(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-getStats
-}
