+++ man/COINToolIn.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{COINToolIn}
-\alias{COINToolIn}
-\title{COINToolIn}
-\usage{
-COINToolIn(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-COINToolIn
-}
