+++ man/iplotIndDist.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{iplotIndDist}
-\alias{iplotIndDist}
-\title{iplotIndDist}
-\usage{
-iplotIndDist(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-iplotIndDist
-}
