+++ man/indDash.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{indDash}
-\alias{indDash}
-\title{indDash}
-\usage{
-indDash(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-indDash
-}
