+++ man/compTableMulti.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{compTableMulti}
-\alias{compTableMulti}
-\title{compTableMulti}
-\usage{
-compTableMulti(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-compTableMulti
-}
