+++ man/iplotCorr.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{iplotCorr}
-\alias{iplotCorr}
-\title{iplotCorr}
-\usage{
-iplotCorr(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-iplotCorr
-}
