+++ man/iplotBar.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{iplotBar}
-\alias{iplotBar}
-\title{iplotBar}
-\usage{
-iplotBar(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-iplotBar
-}
