+++ man/getIn.Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/defunct.R
-\name{getIn}
-\alias{getIn}
-\title{getIn}
-\usage{
-getIn(...)
-}
-\arguments{
-\item{...}{deprecated parameter}
-}
-\description{
-getIn
-}
