### 2022-09-22 ee2403d
doc adjustments following expansion of testing
