+++ man/get_unit_summary.Rd
@@ -15,7 +15,7 @@ get_unit_summary(coin, usel, Levels, dset = "Aggregated", nround = 2)
 
 \item{dset}{The data set within the coin to extract scores and ranks from}
 
-\item{nround}{Number of decimals to round scores to, default 2.}
+\item{nround}{Number of decimals to round scores to, default 2. Set to \code{NULL} to disable rounding.}
 }
 \value{
 A summary table as a data frame, containing scores and ranks for specified indicators/aggregates.
