+++ man/get_stats.Rd
@@ -28,6 +28,6 @@ See also \code{vignette("analysis")}.
 This function replaces the now-defunct \code{getStats()} from COINr < v1.0.
 }
 \examples{
-#
+# see individual method documentation
 
 }
