### 2022-09-22 7789a09
Many unit tests added following JOSS review, see issue #7
