+++ _pkgdown.yml
@@ -143,7 +143,6 @@ reference:
   - check_iMeta
   - check_SkewKurt
   - compare_df
-  - copeland
   - Denominate.coin
   - Denominate.data.frame
   - Denominate.purse
@@ -217,62 +216,10 @@ reference:
   - Treat.purse
   - winsorise
 
-- title: Deprecated functions
-  desc: These are functions in COINr versions before 1.0.0. They now simply redirect to the newer function names. They will be deleted in future updates.
+- title: Data
+  desc: Built in data sets used in examples and testing.
   contents:
   - ASEM_iData
   - ASEM_iData_p
   - ASEM_iMeta
-  - assemble
-  - build_ASEM
-  - checkData
-  - coin2Excel
-  - COINToolIn
-  - coin_win
-  - colourTable
-  - compareDF
-  - compTable
-  - compTableMulti
-  - corrweightscat
-  - effectiveWeight
-  - extractYear
-  - geoMean
-  - geoMean_rescaled
-  - getCorr
-  - getCronbach
-  - getIn
-  - getPCA
-  - getResults
-  - getStats
-  - getStrengthNWeak
-  - getUnitReport
-  - getUnitSummary
-  - harMean
-  - hicorrSP
-  - indChange
-  - indDash
-  - iplotBar
-  - iplotCorr
-  - iplotIndDist
-  - iplotIndDist2
-  - iplotMap
-  - iplotRadar
-  - iplotTable
-  - loggish
-  - names2Codes
-  - noisyWeights
-  - plotCorr
-  - plotframework
-  - plotIndDist
-  - plotIndDot
-  - plotSA
-  - plotSARanks
-  - rankDF
-  - removeElements
-  - replaceDF
-  - resultsDash
-  - rew8r
-  - roundDF
-  - sensitivity
-  - weightOpt
   - WorldDenoms
