### 2022-09-22 bluefoxr/COINr#7 @bluefoxr
Hi @bauer-alex I have added a lot more unit tests and the coverage is now up over 82%. Could you have a look and see if this is sufficient for now, or if not please point to other tests to perform. One untested function you may notice is [import_coin_tool()](https://bluefoxr.github.io/COINr/reference/import_coin_tool.html) which I have left out because it would involve downloading a fairly heavy spreadsheet from the web and I worry about this causing problems with CRAN. Still, if you think this should go ahead please let me know as well. Thanks.
