### 2022-09-22 bluefoxr/COINr#15 @bluefoxr
@paulrougieux do you think the paper modifications are sufficient to close the issue for now? Thanks
