### 2022-09-22 bluefoxr/COINr#7 @bauer-alex
Seems good to me now. :+1:
Did you deliberately not test `COIN_to_coin`? Seems like the only function left for which unit tests would be reasonable.
