### 2022-09-22 bluefoxr/COINr#7 @bluefoxr
It is actually tested but requires another package to be downloaded, perhaps it is not picked up if you don't have that installed. See https://github.com/bluefoxr/COINr/blob/master/tests/testthat/test-convert.R
