### 2022-09-22 bluefoxr/COINr#7 @bauer-alex
Indeed. I see.
Alright, then I'm happy to close the issue.
