### 2022-09-23 bluefoxr/COINr#7 @bluefoxr
Thanks and by the way improving the test coverage helped me uncover a few small bugs, so it has helped to improve the package considerably.
