### 2022-09-23 3464034
Improve Treat() doc
