### 2022-09-23 7cf53b5
Improve Treat() doc re issue #11
