+++ R/treat.R
@@ -4,9 +4,9 @@
 #' details. See also `vignette("treat")`.
 #'
 #' @param x A purse object
-#' @param dset The data set to treat in each coin
-#' @param global_specs Default specifications
-#' @param indiv_specs Individual specifications
+#' @param dset The data set to treat in each coin.
+#' @param global_specs Default specifications. See details in [Treat.coin()].
+#' @param indiv_specs Individual specifications. See details in [Treat.coin()].
 #' @param combine_treat By default, if `f1` fails to pass `f_pass`, then `f2` is applied to the original `x`,
 #' rather than the treated output of `f1`. If `combine_treat = TRUE`, `f2` will instead be applied to the output
 #' of `f1`, so the two treatments will be combined.
@@ -38,10 +38,81 @@ Treat.purse <- function(x, dset, global_specs = NULL, indiv_specs = NULL,
 
 #' Treat a data set in a coin for outliers
 #'
-#' Operates a two-stage data treatment process, based on two data treatment functions, and a pass/fail
-#' function which detects outliers. This function is set up to allow any functions to be passed as the
+#' Operates a two-stage data treatment process on the data set specified by `dset`, based on two data treatment functions, and a pass/fail
+#' function which detects outliers. The method of data treatment can be either specified by the `global_specs` argument (which applies
+#' the same specifications to all indicators in the specified data set), or else (additionally) by the `indiv_specs` argument which allows different
+#' methods to be applied for each indicator. See details. For a simpler function for data treatment, see the wrapper function [qTreat()].
+#'
+#' @details
+#' # Global specifications
+#'
+#' If the same method of data treatment should be applied to all indicators, use the `global_specs` argument. This argument takes a structured
+#' list which looks like this:
+#'
+#' ```
+#' global_specs = list(f1 = .,
+#'                     f1_para = list(.),
+#'                     f2 = .,
+#'                     f2_para = list(.),
+#'                     f_pass = .,
+#'                     f_pass_para = list()
+#'                     )
+#' ```
+#'
+#' The entries in this list correspond to arguments in [Treat.numeric()], and the meanings of each are also described in more detail here
+#' below. In brief, `f1` is the name of a function to apply at the first round of data treatment, `f1_para` is a list of any additional
+#' parameters to pass to `f1`, `f2` and `f2_para` are equivalently the function name and parameters of the second round of data treatment, and
+#' `f_pass` and `f_pass_para` are the function and additional arguments to check for the existence of outliers.
+#'
+#' The default values for `global_specs` are as follows:
+#'
+#' ```
+#' global_specs = list(f1 = "winsorise",
+#'                      f1_para = list(na.rm = TRUE,
+#'                                     winmax = 5,
+#'                                     skew_thresh = 2,
+#'                                     kurt_thresh = 3.5,
+#'                                     force_win = FALSE),
+#'                      f2 = "log_CT",
+#'                      f2_para = list(na.rm = TRUE),
+#'                      f_pass = "check_SkewKurt",
+#'                      f_pass_para = list(na.rm = TRUE,
+#'                                         skew_thresh = 2,
+#'                                         kurt_thresh = 3.5))
+#' ```
+#'
+#' This shows that by default (i.e. if `global_specs` is not specified), each indicator is checked for outliers by the [check_SkewKurt()] function, which
+#' uses skew and kurtosis thresholds as its parameters. Then, if outliers exist, the first function [winsorsise()] is applied, which also
+#' uses skew and kurtosis parameters, as well as a maximum number of winsorised points. If the Winsorisation function does not satisfy
+#' `f_pass`, the [log_CT()] function is invoked.
+#'
[…]
