+++ man/Treat.numeric.Rd
@@ -19,17 +19,17 @@
 \arguments{
 \item{x}{A numeric vector.}
 
-\item{f1}{First stage data treatment function}
+\item{f1}{First stage data treatment function e.g. as a string.}
 
-\item{f1_para}{First stage data treatment function parameters}
+\item{f1_para}{First stage data treatment function parameters as a named list.}
 
-\item{f2}{First stage data treatment function}
+\item{f2}{First stage data treatment function as a string.}
 
-\item{f2_para}{First stage data treatment function parameters}
+\item{f2_para}{First stage data treatment function parameters as a named list.}
 
 \item{f_pass}{A string specifying an outlier detection function - see details. Default \code{"check_SkewKurt"}}
 
-\item{f_pass_para}{Any further arguments to pass to \code{f_pass()}.}
+\item{f_pass_para}{Any further arguments to pass to \code{f_pass()}, as a named list.}
 
 \item{combine_treat}{By default, if \code{f1} fails to pass \code{f_pass}, then \code{f2} is applied to the original \code{x},
 rather than the treated output of \code{f1}. If \code{combine_treat = TRUE}, \code{f2} will instead be applied to the output
