### 2022-09-23 853dc5c
Improvements and fixes to documentation as per issue #11
