### 2022-09-26 97a214b
distance to target normalisation enabled (inc vignette, testing)
