+++ R/normalise.R
@@ -129,6 +129,9 @@ Normalise.purse <- function(x, dset, global_specs = NULL, indiv_specs = NULL,
 #'
 #' Note, all COINr normalisation functions (passed to `f_n`) are of the form `n_*()`. Type `n_` in the R Studio console and press the Tab key to see a list.
 #'
+#' This function includes a special case for "distance to target" normalisation. Setting `global_specs = list(f_n = "n_dist2targ")` will apply distance to
+#' target normalisation, automatically passing targets found in the "Target" column of `iMeta`.
+#'
 #' ## Individual column specification
 #'
 #' Optionally, indicators can be normalised with different normalisation functions and parameters using the
@@ -196,8 +199,41 @@ Normalise.coin <- function(x, dset, global_specs = NULL, indiv_specs = NULL,
 
   # NORMALISE DATA ----------------------------------------------------------
 
-  iData_n <- Normalise(iData_, global_specs = global_specs, indiv_specs = indiv_specs,
-                        directions = dirs_c)
+  if(!is.null(global_specs[["f_n"]])){
+    if(global_specs[["f_n"]] == "n_dist2targ"){
+
+      # special treatment for dist2targ
+      # first, get iMeta
+      iMeta <- coin$Meta$Ind
+      if(is.null(iMeta[["Target"]])){
+        stop("You specified f_para = 'n_dist2targ' but no targets can be found - please attach these as a column 'Target' in iMeta.")
+      }
+      # see if cap_max is specified
+      if(!is.null(global_specs$f_n_para$cap_max)){
+        cap_max <- global_specs$f_n_para$cap_max
+      } else {
+        cap_max <- FALSE
+      }
+
+      # now we need to apply the n_dist2targ() function to each column, but also respecting the directions.
+      l_n <- lapply(names(iData_), function(icode){
+        n_dist2targ(iData_[[icode]],
+                    targ = iMeta$Target[iMeta$iCode == icode],
+                    direction = dirs_c$Direction[dirs_c$iCode == icode],
+                    cap_max = cap_max)
+      })
+      names(l_n) <- names(iData_)
+      iData_n <- as.data.frame(l_n)
+
+    } else {
+      iData_n <- Normalise(iData_, global_specs = global_specs, indiv_specs = indiv_specs,
+                           directions = dirs_c)
+    }
+  } else {
+    iData_n <- Normalise(iData_, global_specs = global_specs, indiv_specs = indiv_specs,
+                         directions = dirs_c)
+  }
+
   # reunite with uCode col
   iData_n <- cbind(uCode = iData$uCode, iData_n)
 
@@ -640,14 +676,23 @@ n_dist2ref <- function(x, iref, cap_max = FALSE){
 
 #' Normalise as distance to target
 #'
-#' A measure of the distance of each value of `x` to a specified target. The formula is:
+#' A measure of the distance of each value of `x` to a specified target which can be a high or low target depending on `direction`. See details below.
 #'
-#' \deqn{ 1 - (x_{targ} - x)/(x_{targ} - x_{min}) }
 #'
-#' Values exceeding `x_targ` can be optionally capped at 1 if `cap_max = TRUE`.
+#' If `direction = 1`, the formula is:
+#'
+#' \deqn{ \frac{x - x_{min}}{x_{targ} - x_{min}} }
+#'
+#' else if `direction = -1`:
+#'
+#' \deqn{ \frac{x_{max} - x}{x_{max} - x_{targ}} }
+#'
+#' Values surpassing `x_targ` in either case can be optionally capped at 1 if `cap_max = TRUE`.
 #'
 #' @param x A numeric vector
 #' @param targ An target value
[…]
