### 2022-09-27 #4567 @bluefoxr
Dear @jbytecode I have addressed all the open issues raised by the reviewers so far. Just waiting for @paulrougieux to see if this issue (https://github.com/bluefoxr/COINr/issues/15) can be closed after some revisions I made to the paper. Not sure if the reviewers want to raise further issues at this stage?
