### 2022-09-30 bluefoxr/COINr#15 @paulrougieux
The short introduction on how to build a coin is a good addition. I see you corrected figure titles, but figure 1 still has an issue with country 3 letters codes being very close to each other, maybe you can change this by changing the figure to a larger figure size in R, this is usually enough to force smaller fonts.
