### 2022-09-30 bluefoxr/COINr#15 @paulrougieux
I looked at the version in this run https://github.com/bluefoxr/COINr/actions/runs/3126882703 All the rest of paper.pdf is good for me.
