### 2022-09-30 bluefoxr/COINr#23 @paulrougieux
Run tests/testthat.R in the actions pipeline
The automated test suite should be run in a Github pipeline /action. I'm not sure this is a requirement of JOSS, but systematically running the test suite at each commit helps catching bugs early. The review guidelines ask for:

>  "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"

Related to the JOSS review https://github.com/openjournals/joss-reviews/issues/4567
