### 2022-09-30 bluefoxr/COINr#23 @paulrougieux
After a fresh install of the development version of COINr

    devtools::install_github("bluefoxr/COINr")

When I run the test suite on my laptop

I get the following error:

```

    "Sub-index", "Index"), quietly = FALSE)`: Entries in iMeta$iCode not found in colnames(iData).
Backtrace:
    ▆
 1. └─COINr::change_ind(coin2, add = "LPI", regen = TRUE) at test-regen.R:49:2
 2.   ├─COINr::Regen(coin, quietly = TRUE)
 3.   └─COINr:::Regen.coin(coin, quietly = TRUE)
 4.     ├─base::suppressMessages(do.call(func, args = f_log))
 5.     │ └─base::withCallingHandlers(...)
 6.     ├─base::do.call(func, args = f_log)
 7.     └─COINr::new_coin(...)
── Error (test-results.R:26:3): unit_summary ───────────────────────────────────
Error in `round(y, decimals)`: invalid second argument of length 0
Backtrace:
    ▆
 1. └─COINr::get_unit_summary(...) at test-results.R:26:2
 2.   └─COINr::round_df(tabout, nround)
 3.     ├─base::data.frame(...)
 4.     └─base::lapply(df, function(y) if (is.numeric(y)) round(y, decimals) else y)
 5.       └─COINr (local) FUN(X[[i]], ...)
── Error (test-trends.R:15:3): approx_df ───────────────────────────────────────
Error in `approx_df(Y, tt, 2012)`: could not find function "approx_df"
── Error (test-trends.R:32:3): CAGR ────────────────────────────────────────────
Error in `CAGR(y, x)`: could not find function "CAGR"
── Error (test-trends.R:45:3): prc_change ──────────────────────────────────────
Error in `prc_change(y, x)`: could not find function "prc_change"
── Error (test-trends.R:58:3): get_trends ──────────────────────────────────────
Error in `get_trends(purse, dset = "Raw")`: could not find function "get_trends"

[ FAIL 15 | WARN 0 | SKIP 1 | PASS 299 ]
Error: Test failures
>

```
