### 2022-09-30 bluefoxr/COINr#23 @paulrougieux
Sorry after a restart (I don't know why a restart was necessary), but anyway, after a restart the test suite is successfull on my laptop and returns the following:

```

Optimisation successful!
[ FAIL 0 | WARN 0 | SKIP 3 | PASS 342 ]

══ Skipped tests ══════════════════════════════════════════════════════════════════════════════
• empty test (1)
• On CRAN (2)

[ FAIL 0 | WARN 0 | SKIP 3 | PASS 342 ]
>

```
