### 2022-09-30 #4567 @paulrougieux
I see 2 issues remaining :
- minor issue https://github.com/bluefoxr/COINr/issues/15  with the country codes in figure one in the paper
- make the test suite run in a Github action to automate them  bluefoxr/COINr#23

Once these are clarified or fixed, the paper is OK from my side.
