### 2022-10-01 06b1ce3
Fixed bug in text colour RE issue #22
