### 2022-10-01 6a95726
Set up github actions for various things
