### 2022-10-01 de9a7d4
Flag COINr6 as dev package
