### 2022-10-01 cb3cb5a
Remove suggests COINr6 and add COIN to data to avoid CRAN check errors with GH package
