+++ man/ASEM_COIN.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{ASEM_COIN}
+\alias{ASEM_COIN}
+\title{ASEM COIN (COINr < v1.0)}
+\format{
+A "COIN" class object
+}
+\source{
+\url{https://github.com/bluefoxr/COINr6}
+}
+\usage{
+ASEM_COIN
+}
+\description{
+This is an "old format" "COIN" object which is stored for testing purposes.
+It is generated using the COINr6 package (only available on GitHub) using
+\code{COINr6::build_ASEM()}
+}
+\keyword{datasets}
