### 2022-10-02 b1eaa14
Doc ASEM_COIN to fix error in pkgdown
