+++ pkgdown.yml
@@ -1,0 +1,26 @@
+pandoc: 2.14.2
+pkgdown: 2.0.6
+pkgdown_sha: ~
+articles:
+  adjustments: adjustments.html
+  aggregate: aggregate.html
+  analysis: analysis.html
+  coins: coins.html
+  data_selection: data_selection.html
+  denomination: denomination.html
+  imputation: imputation.html
+  normalise: normalise.html
+  other_functions: other_functions.html
+  overview: overview.html
+  results: results.html
+  screening: screening.html
+  sensitivity: sensitivity.html
+  treat: treat.html
+  v1: v1.html
+  visualisation: visualisation.html
+  weights: weights.html
+last_built: 2022-10-02T19:19Z
+urls:
+  reference: https://bluefoxr.github.io/COINr/reference
+  article: https://bluefoxr.github.io/COINr/articles
+
