### 2022-10-02 5a2be61
Change plot font to avoid problems on Mac
