+++ R/plot_bar.R
@@ -186,6 +186,7 @@ plot_bar <- function(coin, dset, iCode, ..., uLabel = "uCode", axes_label = "iCo
     plt <- plt + ggplot2::scale_y_log10()
   }
 
-  plt + ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45, hjust = 1))
+  plt + ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45, hjust = 1)) +
+    ggplot2::theme(text=ggplot2::element_text(family="sans"))
 
 }
