+++ R/plot_scatter.R
@@ -172,6 +172,7 @@ plot_scatter <- function(coin, dsets, iCodes, ..., by_group = NULL,
   }
 
 
-  plt
+  plt  +
+    ggplot2::theme(text=ggplot2::element_text(family="sans"))
 
 }
