### 2022-10-02 e750a50
Deploying to gh-pages from @ bluefoxr/COINr@5a2be61eea7938486dfbe1aa847cffb10eba43c9 🚀
