+++ pkgdown.yml
@@ -19,7 +19,7 @@ articles:
   v1: v1.html
   visualisation: visualisation.html
   weights: weights.html
-last_built: 2022-10-02T19:31Z
+last_built: 2022-10-02T20:02Z
 urls:
   reference: https://bluefoxr.github.io/COINr/reference
   article: https://bluefoxr.github.io/COINr/articles
