### 2022-10-02 b219d5c
Deploying to gh-pages from @ bluefoxr/COINr@6ebd02abf9792a6294535ada299c71a1b3d31819 🚀
