### 2022-10-03 bluefoxr/COINr#23 @bluefoxr
Thanks for this comment. I have now added the following GitHub actions:

* [CMD check](https://github.com/bluefoxr/COINr/actions/workflows/R-CMD-check.yaml)
* [Test coverage](https://github.com/bluefoxr/COINr/actions/workflows/test-coverage.yaml)
* [pkgdown site build](https://github.com/bluefoxr/COINr/actions/workflows/pkgdown.yaml)

The CMD check has tests built in and also runs them on different platforms so this seems the most thorough way to automate the testing. I added the other two since they are also useful - update of test coverage and rebuilding the pkgdown site on every commit (previously I was doing this locally). The result of the CMD check is always displayed on the badge at the README page: https://github.com/bluefoxr/COINr

Let me know if this is OK to close the issue. Thanks.
