### 2022-10-03 c94c9d8
smaller x axis font for bar chart
