### 2022-10-03 bluefoxr/COINr#15 @bluefoxr
I have made the font a bit smaller for fig.1 - see https://github.com/bluefoxr/COINr/commit/c94c9d8f9cc65553cab523dcd957d7a184fd1c01 - I don't want to make it too much smaller because it might be hard to read.
