### 2022-10-03 9d96c1f
Deploying to gh-pages from @ bluefoxr/COINr@c94c9d8f9cc65553cab523dcd957d7a184fd1c01 🚀
