### 2022-10-03 #4567 @bluefoxr
Hi @paulrougieux I have made some commits to fix these issues. Please take a look and see if you think they fix the problems. Thanks.
