### 2022-10-03 bluefoxr/COINr#23 @paulrougieux
Looks great. I usually use gitlab CI and I am not knowledgable about Github workflows, where can I see the console output of tests/testthat.R ?
I was trying to look in the workflow called "test coverage" https://github.com/bluefoxr/COINr/actions/runs/3172012270/jobs/5166029329 I see "Run covr::codecov(quiet = FALSE)" but I don't see the output of the actual tests being run. Where should I look for that?
