### 2022-10-03 bluefoxr/COINr#23 @bluefoxr
Yes to be honest I couldn't see how to run only the tests. However the CMD check runs all tests as part of the check, so if a test fails, the CMD check will fail and it will output the error. You can see the output of the CMD check e.g. [here](https://github.com/bluefoxr/COINr/actions/runs/3172012264/jobs/5166029453), and the tests are being run, albeit without a detailed output:

![image](https://user-images.githubusercontent.com/72880276/193527759-159c76f4-63c1-4de0-b251-542da5d02087.png)

Since the CMD check is anyway very useful for the package, would you agree that this is sufficient? I could possibly add another action specifically for the tests, but then that would run the testing twice.
