### 2022-10-03 bluefoxr/COINr#23 @paulrougieux
More details on which test pass or not might be helpful, but I guess R CMD check is good enough to make sure the test suite passes in its entirety. You can close this issue.
