### 2022-10-03 bluefoxr/COINr#23 @bluefoxr
I think that in the event a test doesn't pass, it will output the details of the test and line number etc. Anyway thanks and I will close the issue.
