### 2022-10-03 #4567 @paulrougieux
Both issues  bluefoxr/COINr#23 and  bluefoxr/COINr#15  are fixed, I ticked all boxes in the checklist above.
