### 2022-10-03 #4567 @bluefoxr
> Both issues [bluefoxr/COINr#23](https://github.com/bluefoxr/COINr/issues/23) and [bluefoxr/COINr#15](https://github.com/bluefoxr/COINr/issues/15) are fixed, I ticked all boxes in the checklist above.

Thanks @paulrougieux for your valuable input!
