### 2022-10-05 a03ebd6
Fix colour thresholds bug
