### 2022-10-05 63337aa
Deploying to gh-pages from @ bluefoxr/COINr@a03ebd6bf5f47dfe1bf051292a96d825fcecc996 🚀
