### 2022-10-05 39b4190
Add details of COINr6 package new conversion
