### 2022-10-06 #4567 @bauer-alex
@bluefoxr I now took a deeper look over the paper. Here are my final comments:
First of all, I really like the style of writing. Nice descriptions with a good depth and useful examples and links to further helpful materials. Good work!

Apart from the following long list of (easily solvable) minor points down below, I only have one major point which is that your paper is too long for JOSS. It should include 1.000 words max, but your text parts (without headers, code blocks, figure captions, acknowledgments, references) currently comprise ~1.360 words.
- [ ] Please shorten your paper appropriately, to have ~1.000 words.

Minor points:
- [ ] Please put all your materials for the paper in a dedicated folder `paper`, s.t. the repo is cleaner. Similar to [this example package](https://github.com/bauer-alex/APCtools)
- [ ] In _Existing tools_, first sentence, you use the abbreviation _CIs_. This must be introduced somewhere before using it. E.g. in the very first sentence of the paper: _Composite indicators (CIs) are aggregations ..._
- [ ] In _Existing tools_, second paragraph, you refer to _Excel_. Please write _Microsoft Excel_ to be fully specific.
- [ ] In the following sentence you refer to _Matlab_. I think the company's official writing is _MATLAB_. But that's up to you in my opinion.
- [ ] The Pyhton library _CIF_ is missing a reference.
- [ ] Package names are sometimes written in italic, sometimes bold, sometimes formatted normally. Please be consistent in this. You could write a package name in _italic_ when stating it the very first time, and then without formatting.
- [ ] In the first bullet point of _Why COINr_, don't use _e.g._ and _etc._ simultaneously
- [ ] At the end of _Why COINr_, please cite _ggplot2_ when naming it
- [ ] You refer to the _pkgdown website_ twice in the paper. Once only the word _website_ is a hyperlink, once both words are one hyperlink. Please keep this consistent.
- [ ] The interval _[0,100]_ don't has to be indented like a formula. Just keep it part of the normal text.
- [ ] Correlation plot: I really like the type of visualization! Only thing is, the black text cannot be read on the dark red color. Please a red color that is equally light as the green color.
- [ ] Both figure captions are quite short. Only reading the figure caption should allow for understanding the (visualized information in the) figure, without reading the text. Please edit them accordingly.
