### 2022-10-06 f4eabec
Define CIs acronym
