### 2022-10-06 2014692
Deploying to gh-pages from @ bluefoxr/COINr@3b4a23a877e838c42adccb37169d617a1e2328a0 🚀
