+++ paper/paper.bib
@@ -18,6 +18,15 @@
   doi = "10.13140/RG.2.2.14408.75520"
 }
 
+@online{cif,
+  title={Composite Indicators Framework (CIF) for Business Cycle Analysis},
+  author = {Lenka Vran\'{a}},
+  url={https://pypi.org/project/cif/},
+  urldate = {2022-10-06},
+  year = 2022,
+  publisher={Python Software Foundation}
+}
+
 @article{cinelli2021mcda,
   title={MCDA Index Tool: an interactive software to develop indices and rankings},
   author={Cinelli, Marco and Spada, Matteo and Kim, Wansub and Zhang, Yiwen and Burgherr, Peter},
@@ -141,4 +150,13 @@
   year={2021},
   publisher={Springer},
   doi = "10.1007/s11205-021-02688-6"
+}
+
+@Book{ggplot2,
+  author = {Hadley Wickham},
+  title = {ggplot2: Elegant Graphics for Data Analysis},
+  publisher = {Springer-Verlag New York},
+  year = {2016},
+  isbn = {978-3-319-24277-4},
+  url = {https://ggplot2.tidyverse.org},
 }
