### 2022-10-06 b83d57e
Increase fig size slightly
