### 2022-10-06 fb9fb5e
Deploying to gh-pages from @ bluefoxr/COINr@b83d57eff756240c2d40da59aa5fb6025d91039e 🚀
