### 2022-10-06 b9d0f56
Deploying to gh-pages from @ bluefoxr/COINr@2c7275da2cd75780d78ac3ab79523a2dd197c66a 🚀
