### 2022-10-06 bc91efe
Further slimming down
