+++ paper/paper.md
@@ -51,9 +51,9 @@ The Python library *CIF* gives a number of tools for building composite indicato
 
 COINr is a significant step beyond existing composite indicator tools in many respects. COINr wraps all composite indicator data, analysis and methodological choices into a single S3 class object called a "coin". A coin is a structured list including:
 
-* Indicator data sets for each processing step (e.g. imputation and normalisation)
-* Metadata pertaining to indicators and units (e.g. names and weights, but also the hierarchical structure of the index)
-* A record of the COINr functions applied in constructing the coin
+-   Indicator data sets for each processing step (e.g. imputation and normalisation)
+-   Metadata pertaining to indicators and units (e.g. names and weights, but also the hierarchical structure of the index)
+-   A record of the COINr functions applied in constructing the coin
 
 This enables a neat and structured environment, simplifies the syntax of functions, and also allows comparisons between different versions of the same index, as well as global sensitivity analysis along the lines of [@saisana2005uncertainty] (for the distinction between "local" and "global" sensitivity analysis, see e.g. [@saltelli2019so]). COINr also supports time-indexed (panel) data, represented by the "purse" class (a data frame containing a time-indexed collection of coins). For more information on coins and purses, see the "coins" [vignette](https://bluefoxr.github.io/COINr/articles/coins.html).
 
@@ -67,61 +67,52 @@ Primarily, COINr is used for *building* composite indicators: in practice this w
 
 To give a flavour of COINr, we give a very short example using the built-in "ASEM" data set which comprises two data frames (one of indicator data, and the other of metadata). To build a coin, we call `new_coin()`:
 
-```
-# load COINr
-library(COINr)
+    # load COINr
+    library(COINr)
 
-# build a coin with example data set
-coin <- new_coin(iData = ASEM_iData, iMeta = ASEM_iMeta)
-```
+    # build a coin with example data set
+    coin <- new_coin(iData = ASEM_iData, iMeta = ASEM_iMeta)
 
 To see how these data frames are formatted, use e.g. `str(ASEM_iData)` or `View(ASEM_iData)` and see the "coins" [vignette](https://bluefoxr.github.io/COINr/articles/coins.html).
 
 In the most simple case, we could build a composite indicator by [normalising](https://bluefoxr.github.io/COINr/articles/normalise.html) the indicators (bringing them onto a common scale), and [aggregating](https://bluefoxr.github.io/COINr/articles/aggregate.html) them (using weighted averages to calculate index scores). This can be done in COINr using the `Normalise()` and `Aggregate()` functions respectively:
 
-```
-# normalise (scale) each indicator onto [0, 100] interval
-coin <- qNormalise(coin, dset = "Raw", f_n = "n_minmax",
-                   f_n_para = list(l_u = c(0, 100)))
-                   
-# aggregate using weighted arithmetic mean
-# (note weights are input in data frames when calling new_coin() )
-coin <- Aggregate(coin, dset = "Normalised", f_ag = "a_amean") 
-```
+    # normalise (scale) each indicator onto [0, 100] interval
+    coin <- qNormalise(coin, dset = "Raw", f_n = "n_minmax",
+                       f_n_para = list(l_u = c(0, 100)))
+                       
+    # aggregate using weighted arithmetic mean
+    # (note weights are input in data frames when calling new_coin() )
+    coin <- Aggregate(coin, dset = "Normalised", f_ag = "a_amean") 
+
[…]
