### 2022-10-06 38a0d46
Deploying to gh-pages from @ bluefoxr/COINr@bc91efef554fb5e8130398adec735ca4184b90f9 🚀
