### 2022-10-06 #4567 @bauer-alex
Thanks for the quick changes @bluefoxr.
For the word count I get 1,097 words (again, excluding header, figure captions, references and acknowledgements). Seeing that - as you mention - the paper grew a little during the revision process, this is fine for me.

The only open task from my previous message is to name it _Microsoft Excel_ instead of only _Excel_. Then you get my go for publication.
