### 2022-10-07 faf9b00
Deploying to gh-pages from @ bluefoxr/COINr@b0c33138bdb56fbc66fe761dfc0ee57efa677d2a 🚀
