### 2022-10-07 #4567 @bluefoxr
Thanks also for the quick reply, I have just made the last change - somehow I overlooked that one: https://github.com/bluefoxr/COINr/commit/b0c33138bdb56fbc66fe761dfc0ee57efa677d2a
