### 2022-10-07 9107adf
Deploying to gh-pages from @ bluefoxr/COINr@f0198b6eebf2007c452dcfc2d30a47147a77360a 🚀
