### 2022-10-07 fa5ca26
update paper and bibtex
