+++ paper/paper.bib
@@ -162,3 +162,13 @@
   isbn = {978-3-319-24277-4},
   url = {https://ggplot2.tidyverse.org},
 }
+
+
+@Manual{rstats,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2022},
+    url = {https://www.R-project.org/},
+  }
