### 2022-10-07 #4567 @bluefoxr
@jbytecode thanks for your input - I have merged that as well.
