### 2022-10-07 d2fffe2
Deploying to gh-pages from @ bluefoxr/COINr@712d3ee7c9bfa64ae2548143edf16976d69168bf 🚀
