+++ paper/paper.md
@@ -51,9 +51,9 @@ The Python library *CIF* gives a number of tools for building composite indicato
 
 COINr is a significant step beyond existing composite indicator tools in many respects. COINr wraps all composite indicator data, analysis and methodological choices into a single S3 class object called a "coin". A coin is a structured list including:
 
--   Indicator data sets for each processing step (e.g. imputation and normalisation)
--   Metadata pertaining to indicators and units (e.g. names and weights, but also the hierarchical structure of the index)
--   A record of the COINr functions applied in constructing the coin
+-   Indicator data sets for each processing step (e.g. imputation and normalisation).
+-   Metadata pertaining to indicators and units (e.g. names and weights, but also the hierarchical structure of the index).
+-   A record of the COINr functions applied in constructing the coin.
 
 This enables a neat and structured environment, simplifies the syntax of functions, and also allows comparisons between different versions of the same index, as well as global sensitivity analysis along the lines of [@saisana2005uncertainty] (for the distinction between "local" and "global" sensitivity analysis, see e.g. [@saltelli2019so]). COINr also supports time-indexed (panel) data, represented by the "purse" class (a data frame containing a time-indexed collection of coins). For more information on coins and purses, see the "coins" [vignette](https://bluefoxr.github.io/COINr/articles/coins.html).
 
@@ -65,7 +65,7 @@ COINr also offers a far wider range of functions and methodological options than
 
 Primarily, COINr is used for *building* composite indicators: In practice this would usually involve assembling a set of indicators (usually from different sources) and accompanying metadata, and assembling them into data frames that can be read by COINr to build a "coin" (see [vignette](https://bluefoxr.github.io/COINr/articles/coins.html)). After that, the composite scores are calculated by operating on the coin using any of the "building functions", which specify the methodological steps to apply, and *how* to apply them.
 
-To give a flavour of COINr, we present a very short example using the built-in "ASEM" data set which comprises two data frames (one of indicator data, and the other of metadata). To build a coin, we call `new_coin()`:
+To give a flavour of COINr, we present a very short example using the built-in "ASEM" data set which comprises two data frames (one of indicator data, and the other of metadata). To build a coin, `new_coin()` function is called:
 
 ```R
 # load COINr
@@ -75,7 +75,7 @@ library(COINr)
 coin <- new_coin(iData = ASEM_iData, iMeta = ASEM_iMeta)
 ```
 
-To see how these data frames are formatted, use e.g. `str(ASEM_iData)` or `View(ASEM_iData)` and see the "coins" [vignette](https://bluefoxr.github.io/COINr/articles/coins.html).
+To see how these data frames are formatted, one can use e.g. `str(ASEM_iData)` or `View(ASEM_iData)` and see the "coins" [vignette](https://bluefoxr.github.io/COINr/articles/coins.html).
 
 In the most simple case, we could build a composite indicator by [normalising](https://bluefoxr.github.io/COINr/articles/normalise.html) the indicators (bringing them onto a common scale), and [aggregating](https://bluefoxr.github.io/COINr/articles/aggregate.html) them (using weighted averages to calculate index scores). This can be done in COINr using the `Normalise()` and `Aggregate()` functions respectively:
 
@@ -89,9 +89,9 @@ coin <- qNormalise(coin, dset = "Raw", f_n = "n_minmax",
 coin <- Aggregate(coin, dset = "Normalised", f_ag = "a_amean") 
 ```
 
-Both of these functions allow any other function to be passed to them, allowing more complex types of normalisation and aggregation. Here, we have simply used the "min-max" normalisation method (scaling indicators onto the $[0, 100]$ interval), and aggregated using the weighted arithmetic mean.
[…]
