### 2022-10-07 #4567 @jbytecode
Dear @bluefoxr

- [X] The paper and the bibtex look nicer now. Thank you.
- [ ] Please read the paper pdf carefully, one more time, correct any issue if exists. Please perform a full grammar check.
- [ ] Create a tagged release in the software repository. Report the version here. The version should be in form of `vx.y.z`, e.g. `v1.2.3`.
- [ ] Create a Zenodo archive of the repository in https://zenodo.org/
- [ ] The Zenodo archive (with the same name/title with the JOSS paper) should have the correct metadata, e.g. **author names**, **ORCID**s, the title of the archive. **Please set the archive title as the paper title.**
- [ ] Please provide the tagged release id, zenodo archive DOI, and Zenodo archive URL here.

Thank you in advance.
