### 2022-10-10 30e7e4b
Final paper adjustments
