### 2022-10-10 #4567 @bluefoxr
@jbytecode I am aiming to complete your checklist by the end of the day, if not by tomorrow. I just made some minor last edits to the paper and will now proceed to tag and archive the repo as requested.
