### 2022-10-10 beaa064
Bump to 1.1, version for JOSS paper
