### 2022-10-10 #4567 @bluefoxr
@jbytecode could I ask which zenodo community, if any, the software should be archived in? Am I free to choose or is there a JOSS policy/requirement? Thanks.
