### 2022-10-10 #4567 @jbytecode
@bluefoxr - The latest snapshot should be registered as a tagged archive in github and should be also registered on an outer service such as Zenodo.org. Yes, this process is required in JOSS submissions. Thank you in advance.
