### 2022-10-10 #4567 @jbytecode
@bluefoxr - Please take a look at the https://joss.readthedocs.io/en/latest/submitting.html?highlight=zenodo#the-review-process
