### 2022-10-10 #4567 @bluefoxr
@jbytecode I have now archived the package on Zenodo. Details are:

* GitHub tagged release: https://github.com/bluefoxr/COINr/releases/tag/v1.1.0
* Zenodo DOI: https://doi.org/10.5281/zenodo.7180388
* Zenodo link: https://zenodo.org/record/7180389#.Y0PzBnZBzN8

Please let me know if it is OK like this or I need to change something. Thanks.
