### 2022-10-10 #4567 @bluefoxr
@jbytecode I can't see anything missing - only three of the authors have ORCIDs, and it corresponds to the JOSS paper as far as I can see? Is there something specific missing?
