### 2022-10-10 #4567 @editorialbot
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
