### 2022-10-10 #4567 @bluefoxr
@jbytecode @bauer-alex @paulrougieux Can I also extend a big thank you for taking your valuable time to review the paper and the package. I would also like to add - I really like the JOSS review system through GitHub: transparent and efficient! Hopefully other journals might start moving in this direction! :)
