### 2022-10-10 #4567 @kthyng
Hi @bluefoxr! Small comments on your paper:
* at least one place has a parenthetical reference when it should be in line (2nd page)
* some proper nouns need to be capitalized in your references. You can preserve or add capitalization in your .bib file by placing {} around characters or words.
