+++ paper/paper.md
@@ -55,7 +55,7 @@ COINr is a significant step beyond existing composite indicator tools in many re
 -   Metadata pertaining to indicators and units (e.g. names and weights, but also the hierarchical structure of the index).
 -   A record of the COINr functions applied in constructing the coin.
 
-This enables a neat and structured environment, simplifies the syntax of functions, and also allows comparisons between different versions of the same index, as well as global sensitivity analysis along the lines of [@saisana2005uncertainty] (for the distinction between "local" and "global" sensitivity analysis, see e.g. [@saltelli2019so]). COINr also supports time-indexed (panel) data, represented by the "purse" class (a data frame containing a time-indexed collection of coins). For more information on coins and purses, see the [coins vignette](https://bluefoxr.github.io/COINr/articles/coins.html).
+This enables a neat and structured environment, simplifies the syntax of functions, and also allows comparisons between different versions of the same index, as well as global sensitivity analysis along the lines of `@saisana2005uncertainty` (for the distinction between "local" and "global" sensitivity analysis, see e.g. `@saltelli2019so`). COINr also supports time-indexed (panel) data, represented by the "purse" class (a data frame containing a time-indexed collection of coins). For more information on coins and purses, see the [coins vignette](https://bluefoxr.github.io/COINr/articles/coins.html).
 
 All major COINr functions have methods for coins, and many have methods for purses, data frames, and numerical vectors. This means that COINr can be used either as an integrated development environment via coins and purses, but equally as a toolbox of functions for other related purposes.
 
