### 2022-10-11 #4567 @jbytecode
@bluefoxr - I've just sent a pull request that correct (hopefully) our editor in chief's suggestions.  Please check them. If you are agreed with them, please apply.

PR: https://github.com/bluefoxr/COINr/pull/28

Thank you in advance.
