+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @manual{CIAOtool,
-      title        = "Composite Indicator Analysis and Optimization ({CIAO}) Tool, v.2",
+      title        = "{C}omposite {I}ndicator {A}nalysis and {O}ptimization ({CIAO}) Tool, v.2",
       author       = "Lindén, D. and Cinelli, M. and Spada, M. and Becker, W. and Burgherr, P.",
       year         = 2021,
       url = {http://dx.doi.org/10.13140/RG.2.2.14408.75520},
@@ -19,7 +19,7 @@
 }
 
 @online{cif,
-  title={Composite Indicators Framework (CIF) for Business Cycle Analysis},
+  title={{C}omposite {I}ndicators {F}ramework (CIF) for Business Cycle Analysis},
   author = {Lenka Vran\'{a}},
   url={https://pypi.org/project/cif/},
   urldate = {2022-10-06},
@@ -88,7 +88,7 @@
 }
 
 @article{cobham2015financial,
-  title={The financial secrecy index: Shedding new light on the geography of secrecy},
+  title={The {F}inancial {S}ecrecy {I}ndex: Shedding new light on the geography of secrecy},
   author={Cobham, Alex and Jansk{\`y}, Petr and Meinzer, Markus},
   journal={Economic geography},
   volume={91},
@@ -100,7 +100,7 @@
 }
 
 @book{HDI2020,
-	title = {The 2020 Human Development Report},
+	title = {The 2020 {H}uman {D}evelopment {R}eport},
 	isbn = {978-92-1-126442-5},
 	language = {English},
 	author = {Conceição, P.},
@@ -143,7 +143,7 @@
 }
 
 @article{stefana2021composite,
-  title={Composite indicators to measure quality of working life in Europe: a systematic review},
+  title={Composite indicators to measure quality of working life in {E}urope: a systematic review},
   author={Stefana, Elena and Marciano, Filippo and Rossi, Diana and Cocca, Paola and Tomasoni, Giuseppe},
   journal={Social Indicators Research},
   volume={157},
