### 2022-10-11 #4567 @bluefoxr
@jbytecode thank you for the edits. I added an extra couple of things as well just a minute ago after the merge. @kthyng thank you for looking over the paper, I think the problems you mentioned are fixed now.
