### 2022-10-11 #4567 @bluefoxr
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04567/joss.04567/10.21105.joss.04567.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.04567/joss.04567/10.21105.joss.04567.pdf) 📄 👈

I checked and this version includes my latest edits
