### 2022-10-11 f25c183
Deploying to gh-pages from @ bluefoxr/COINr@f3d01e05ced036a5540e55417ef340cd5db18edb 🚀
