### 2022-10-11 #4567 @bluefoxr
> Congratulations on your new publication @bluefoxr! Many thanks to editor @jbytecode and reviewers @bauer-alex and @paulrougieux for your time, hard work, and expertise!!

Thank you all again for your time! Very impressed by the publication process at JOSS! All the best to you all.
